Change demo link (error 404)
The demo link redirects to a 404 error.
I don't know if the link to "demo" makes sense because it redirects to the instance search page. And this link already exists in the "Documentation and resources" part
Maybe we should put the link to the main instance as demo? https://lemmy.ml/


See:
https://github.com/YunoHost-Apps/lemmy_ynh/pull/10#issuecomment-933294506